+++ SPA_Symfony-Angular/backend/src/Controller/API/FoodController.php
@@ -33,10 +33,16 @@ class FoodController extends AbstractController
      * @param FoodRepository $foodRepository
      * @return Response
      */
-    public function index(Request $request, FoodRepository $foodRepository): Response
+    public function index(Request $request, FoodRepository $foodRepository, FloorRepository $floorRepository): Response
     {
         $id_floor = $request->query->get('idFloor');
         try {
+            $floor = $floorRepository->findOneById($id_floor);
+            if (!$floor) {
+                return $this->json([
+                    'errors' => "unable to fetch foods at this moment."
+                ], 400);
+            }
             $listFood = $foodRepository->findByIdFloor($id_floor);
 
             $defaultContext = [
@@ -44,7 +50,7 @@ class FoodController extends AbstractController
                     return $object->getId();
                 },
                 ObjectNormalizer::CIRCULAR_REFERENCE_LIMIT =>0,
-                AbstractNormalizer::IGNORED_ATTRIBUTES =>['fridge', 'user', 'floor', 'idFridge'],
+                AbstractNormalizer::IGNORED_ATTRIBUTES =>['fridge', 'user', 'floor', 'idFridge', "qtyFood"],
                 ObjectNormalizer::ENABLE_MAX_DEPTH => true,
             ];
 
@@ -59,7 +65,7 @@ class FoodController extends AbstractController
 
         } catch (\Exception $exception) {
             return $this->json([
-                'errors' => $exception
+                'errors' => "unable to fetch foods at this moment."
             ], 400);
         }
     }
@@ -69,34 +75,40 @@ class FoodController extends AbstractController
      */
     public function newFood(Request $request, FloorRepository $floorRepository): Response
     {
-        $food = new Food();
-
-        $data = json_decode($request->getContent(), true);
-        $name = $data['name'];
-        $type = $data['type'];
-        $expirationDate = $data['expiration_date'];
-        $quantity = $data['quantity'];
-        $dateOfPurchase = $data['date_of_purchase'];
-        $imageFood = $data['image_food_path'];
-        $unitQuantity = $data['unit_qty'];
-        $id_floor = $data['id_floor'];
-
-        $floor = $floorRepository->findOneById($id_floor);
-
-        $food->setName($name);
-        $food->setType($type);
-        $food->setExpirationDate($expirationDate);
-        $food->setQuantity($quantity);
-        $food->setDateOfPurchase($dateOfPurchase);
-        $food->setImageFoodPath($imageFood);
-        $food->setUnitQty($unitQuantity);
-        $food->setIdFloor($floor);
-
-
-        $nbr_floors = $floor->getQtyFood();
-        $floor->setQtyFood($nbr_floors + 1);
-
         try {
+            $food = new Food();
+
+            $data = json_decode($request->getContent(), true);
+            $name = $data['name'];
+            $type = $data['type'];
+            $expirationDate = $data['expiration_date'];
+            $quantity = $data['quantity'];
+            $dateOfPurchase = $data['date_of_purchase'];
+            $imageFood = $data['image_food_path'];
+            $unitQuantity = $data['unit_qty'];
+            $id_floor = $data['id_floor'];
+
+            $floor = $floorRepository->findOneById($id_floor);
+
+            if (!$floor) {
+                return $this->json([
+                    'errors' => "unable to save new food at this moment."
+                ], 400);
+            }
+
+            $food->setName($name);
+            $food->setType($type);
+            $food->setExpirationDate($expirationDate);
+            $food->setQuantity($quantity);
+            $food->setDateOfPurchase($dateOfPurchase);
+            $food->setImageFoodPath($imageFood);
+            $food->setUnitQty($unitQuantity);
+            $food->setIdFloor($floor);
+
+
+            $nbr_floors = $floor->getQtyFood();
+            $floor->setQtyFood($nbr_floors + 1);
+
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($food);
             $entityManager->flush();
@@ -110,64 +122,39 @@ class FoodController extends AbstractController
         }
     }
 
-    public function uploadImage(FormInterface $form, Food $food, $noImageUpload)
-    {
-        $ImageFile = $form->get('imageFood')->getData();
-
-        // this condition is needed because the 'image' field is not required
-        // so the PDF file must be processed only when a file is uploaded
-        if ($ImageFile) {
-            $originalFilename = pathinfo($ImageFile->getClientOriginalName(), PATHINFO_FILENAME);
-            // this is needed to safely include the file name as part of the URL
-            $safeFilename = transliterator_transliterate('Any-Latin; Latin-ASCII; [^A-Za-z0-9_] remove; Lower()', $originalFilename);
-            $newFilename = $safeFilename . '-' . uniqid() . '.' . $ImageFile->guessExtension();
-
-            // Move the file to the directory where brochures are stored
-            try {
-                $ImageFile->move(
-                    'images/fridge',
-                    $newFilename
-                );
-            } catch (FileException $e) {
-                // ... handle exception if something happens during file upload
-            }
-
-            $filePath = 'images/fridge' . '/' . $newFilename;
-            // updates the 'brochureFilename' property to store the PDF file name
-            // instead of its contents
-            $food->setImageFoodPath($filePath);
-        } else {
-            $food->setImageFoodPath($noImageUpload);
-        }
-    }
-
     /**
      * @Route("/{id}", name="food_edit", methods={"PUT"})
      */
     public function edit(Request $request, FoodRepository $foodRepository): Response
     {
-        $id_food = $request->attributes->get('id');
-        $food = $foodRepository->findOneById($id_food);
-
-        $data = json_decode($request->getContent(), true);
-
-        $name = $data['name'];
-        $type = $data['type'];
-        $expirationDate = $data['expiration_date'];
-        $quantity = $data['quantity'];
-        $dateOfPurchase = $data['date_of_purchase'];
-        $imageFoodPath = $data['image_food_path'];
-        $unitQty = $data['unit_qty'];
-
-        $food->setName($name);
-        $food->setType($type);
-        $food->setExpirationDate($expirationDate);
-        $food->setQuantity($quantity);
-        $food->setDateOfPurchase($dateOfPurchase);
-        $food->setImageFoodPath($imageFoodPath);
-        $food->setUnitQty($unitQty);
-
         try {
+            $id_food = $request->attributes->get('id');
+            $food = $foodRepository->findOneById($id_food);
+
+            if (!$food) {
+                return $this->json([
+                    'errors' => "unable to edit food at this moment."
+                ], 400);
+            }
+
+            $data = json_decode($request->getContent(), true);
+
+            $name = $data['name'];
+            $type = $data['type'];
+            $expirationDate = $data['expiration_date'];
+            $quantity = $data['quantity'];
+            $dateOfPurchase = $data['date_of_purchase'];
+            $imageFoodPath = $data['image_food_path'];
+            $unitQty = $data['unit_qty'];
+
+            $food->setName($name);
+            $food->setType($type);
+            $food->setExpirationDate($expirationDate);
+            $food->setQuantity($quantity);
+            $food->setDateOfPurchase($dateOfPurchase);
+            $food->setImageFoodPath($imageFoodPath);
+            $food->setUnitQty($unitQty);
+
             $this->getDoctrine()->getManager()->flush();
 
             return $this->json([
@@ -176,7 +163,7 @@ class FoodController extends AbstractController
 
         } catch (\Exception $exception) {
             return $this->json([
-                'errors' => $exception
+                'errors' => "unable to edit food at this moment."
             ], 400);
         }
     }
@@ -186,15 +173,21 @@ class FoodController extends AbstractController
      */
     public function delete(Request $request, FoodRepository $foodRepository, FloorRepository $floorRepository): Response
     {
-        $id_food = $request->attributes->get('id');
-        $food = $foodRepository->findOneById($id_food);
-
-        $floor = $floorRepository->findOneById($food->getIdFloor());
-
-        $nbr_floors = $floor->getQtyFood();
-        $floor->setQtyFood($nbr_floors - 1);
-
         try {
+            $id_food = $request->attributes->get('id');
+            $food = $foodRepository->findOneById($id_food);
+
+            if (!$food) {
+                return $this->json([
+                    'errors' => "unable to delete food at this moment."
+                ], 400);
+            }
+
+            $floor = $floorRepository->findOneById($food->getIdFloor());
+
+            $nbr_floors = $floor->getQtyFood();
+            $floor->setQtyFood($nbr_floors - 1);
+
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($food);
             $entityManager->flush();
